test: move configuration mock to helper function
