+++ src/sp-proxy/src/frameworks-drivers/main/routes/metadataRouter.test.ts
@@ -1,11 +1,9 @@
 import routes from '@sp-proxy/frameworks-drivers/main/routes'
 import { InvalidRequestError } from '@sp-proxy/interface-adapters/delivery/errors/InvalidRequestError'
 import { GenerateMetadataController } from '@sp-proxy/interface-adapters/delivery/GenerateMetadataController'
-import * as configRepo from '@sp-proxy/interface-adapters/data/FileReadProxyConfig'
 import express from 'express'
 import request from 'supertest'
-import { SpProxyConfigProps } from '@sp-proxy/entities/protocols/SpProxyConfigProps'
-import { SpProxyConfig } from '@sp-proxy/entities/SpProxyConfig'
+import { mockSpProxyConfig } from '@sp-proxy/frameworks-drivers/main/mocks/mockSpProxyConfig.mock'
 jest.mock('@sp-proxy/interface-adapters/data/FileReadProxyConfig')
 
 describe('metadataRoute', () => {
@@ -13,27 +11,7 @@ describe('metadataRoute', () => {
   beforeAll(async () => {
     // setup app for testing this route
     app.use(routes)
-    const mockedProps: SpProxyConfigProps = {
-      host: 'myhost.com',
-      requestedIdentifierFormat: 'my:name:identifier:requested',
-      authnContextIdentifierFormat: 'my:authn:name:identifier:format',
-      skipRequestCompression: true,
-      decryption: {
-        publicCertPath: process.cwd() + '/src/testdata/cert.pem',
-        privateKeyPath: process.cwd() + '/src/testdata/key.pem'
-      },
-      signing: {
-        publicCertPath: process.cwd() + '/src/testdata/cert.pem',
-        privateKeyPath: process.cwd() + '/src/testdata/key.pem'
-      }
-    }
-    const mockedConfig: SpProxyConfig = new SpProxyConfig(mockedProps)
-    jest
-      .spyOn(configRepo.FileReadProxyConfig.prototype, 'read')
-      .mockResolvedValueOnce(mockedConfig)
-
-    // cfg.database.file.proxyConfigPath =
-    //   process.cwd() + '/src/testdata/sp-proxy-config-test.json'
+    mockSpProxyConfig()
   })
   it('should return 400 if controller throws InvalidRequestError', async () => {
     jest
